[DF-10434] Build fix
Former-commit-id: 81c4487dab3dbb8bbac217c5d8c0fae4f0fd1546

+++ Sources/Internal/UI/Private/CEFController.cpp
@@ -178,7 +178,7 @@ void CEFControllerImpl::CleanCache()
     FilePath cachePath = GetCachePath();
     Vector<FilePath> cacheDirContent = fs->EnumerateFilesInDirectory(cachePath);
 
-    uint32 cacheSize = 0;
+    size_t cacheSize = 0;
     for (const FilePath& path : cacheDirContent)
     {
         uint32 size = 0;
